refactor: extract status command handler
